# src/main/resources/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ethereum Enterprise API",
        "version": "1.0.0",
        "title": "Ethereum Enterprise",
        "contact": {}
    },
    "host": "10.196.144.225",
    "schemes": [
        "https"
    ],
    "basePath": "/ethereum-enterprise/api",
...
    "paths": {
        "/block/get": {
            "post": {
                "tags": [
                    "block"
                ],
                "description": "Retrieve a block by number, hash or tag",
                "operationId": "postBlock",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "number": {
                                    "description": "Block number to retrieve",
                                    "type": "integer",
                                    "required": false
                                },
                                "hash": {
                                    "description": "Block hash to retrieve",
                                    "type": "string",
                                    "required": false
                                },
                                "tag": {
                                    "description": "\"earliest\", \"latest\" or \"pending\"",
                                    "type": "string",
                                    "required": false
                                }
                            }
                        }
                    }
                ],
                "examples": [
                    {
                        "name": "Get block by number",
                        "parameters": {
                            "body": "{\n  \"number\": 0\n}\n"
                        }
                    },
                    {
                        "name": "Get block by hash",
                        "parameters": {
                            "body": "{\n  \"hash\": \"0x8564f939768d96f6fc0ba1334ed083ab6538da76f17b6d264082cb69aadc7b4c\"\n}\n"
                        }
                    },
                    {
                        "name": "Get latest block",
                        "parameters": {
                            "body": "{\n  \"tag\": \"latest\"\n}\n"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "description": "response data",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "Block hash identifier"
                                        },
                                        "type": {
                                            "type": "string",
                                            "description": "resource type identifier, 'block'"
                                        },
                                        "attributes": {
                                            "$ref": "#/definitions/Block"
                                        }
                                    }
                                },
                                "errors": {
                                    "description": "Only present if errors were found",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/APIError"
                                    }
                                },
                                "meta": {
                                    "type": "object",
                                    "description": "Extra metadata, such as API version"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/contract/get": {
            "post": {
                "tags": [
                    "contract"
                ],
                "operationId": "postContractGet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "description": "ID (address) of Contract to retrieve",
                                    "example": "0x01f999deef030a0f3c55cec46e6693c06c6b01f8"
                                }
                            }
                        }
                    }
                ],
                "examples": [
                    {
                        "name": "Create SimpleStorage contract",
                        "parameters": {
                            "body": "{\n  \"address\": \"<ENTER ADDRESS HERE>\"\n}\n"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "description": "response data",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "Contract ID"
                                        },
                                        "type": {
                                            "type": "string",
                                            "description": "resource type identifier, 'contract'"
                                        },
                                        "attributes": {
                                            "$ref": "#/definitions/Contract"
                                        }
                                    }
                                },
                                "errors": {
                                    "description": "Only present if errors were found",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/APIError"
                                    }
                                },
                                "meta": {
                                    "type": "object",
                                    "description": "Extra metadata, such as API version"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/transaction/get": {
            "post": {
                "tags": [
                    "transaction"
                ],
                "description": "Retrieve a transaction by txid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "sender": {
                                    "type": "string",
                                    "required": true,
                                    "description": "Request sender pubkey"
                                },
                                "id": {
                                    "type": "string",
                                    "required": true,
                                    "description": "Transaction ID (address) to retrieve",
                                    "example": ""
                                }
                            }
                        }
                    }
                ],
                "examples": [
                    {
                        "name": "Retrieve Transaction information",
                        "parameters": {
                            "body": "{\n  \"address\": \"<ENTER ADDRESS HERE>\"\n}\n"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "description": "response data",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "Transaction ID"
                                        },
                                        "type": {
                                            "type": "string",
                                            "description": "resource type identifier, 'transaction'"
                                        },
                                        "attributes": {
                                            "$ref": "#/definitions/Transaction"
                                        }
                                    }
                                },
                                "errors": {
                                    "description": "Only present if errors were found",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/APIError"
                                    }
                                },
                                "meta": {
                                    "type": "object",
                                    "description": "Extra metadata, such as API version"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/node-control/start": {
            "post": {
                "tags": [
                    "node"
                ],
                "description": "Start ethereum node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": null
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/node-control/stop": {
            "post": {
                "tags": [
                    "node"
                ],
                "description": "Stop ethereum node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "properties": null
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "APIError": {
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Error ID, if available"
                },
                "status": {
                    "type": "string",
                    "description": "HTTP status code"
                },
                "code": {
                    "type": "string",
                    "description": "API-specific error code, if available"
                },
                "title": {
                    "type": "string",
                    "description": "Short error description"
                },
                "detail": {
                    "type": "string",
                    "description": "Detailed error information"
                }
            }
        },
        "Contract": {
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Contract unique identifier. In Ethereum this is the contract address."
                },
                "author": {
                    "type": "string",
                    "description": "Pubkey of sender on original create."
                },
                "keys": {
                    "type": "array",
                    "description": "Pubkeys that can access contract.",
                    "items": {
                        "type": "string"
                    }
                },
                "abi": {
                    "type": "string",
                    "description": "Contract ABI"
                },
                "code": {
                    "type": "string",
                    "description": "Contract source code"
                },
                "code_type": {
                    "type": "string",
                    "description": "Type of code being submitted",
                    "default": "solidity",
                    "enum": [
                        "solidity",
                        "binary",
                        "llvm",
                        "serpent"
                    ]
                }
            }
        },
...
        "Transaction": {
            "properties": {
                "id": {
                    "type": "string"
                },
                "blockHash": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "integer"
                },
                "transactionIndex": {
                    "type": "integer",
                    "description": "integer of the transaction's index position in the block"
                },
                "from": {
                    "type": "string",
                    "description": "address of the sender"
                },
                "to": {
                    "type": "string",
                    "description": "address of the receiver. null when it is a contract creation transaction."
                },
                "cumulativeGasUsed": {
                    "type": "integer",
                    "description": "the total amount of gas used when this transaction was executed in the block."
                },
                "gasUsed": {
                    "type": "integer",
                    "description": "the amount of gas used by this specific transaction alone."
                },
                "contractAddress": {
                    "type": "string",
                    "description": "the contract address created, if this transaction was a contract creation, otherwise null."
                },
                "logs": {
                    "type": "array",
                    "description": "TODO add Log type",
                    "items": {
                        "type": "object"
                    }
                }
            }
        },
        "Block": {
            "properties": {
                "number": {
                    "type": "integer",
                    "format": "int32",
                    "description": "Block number"
                },
                "hash": {
                    "type": "string",
                    "description": "hash of the block"
                },
                "parentHash": {
                    "type": "string",
                    "description": "hash of the parent block"
                },
                "nonce": {
                    "type": "string",
                    "description": "hash of the generated proof-of-work (if avail)"
                },
                "sha3Uncles": {
                    "type": "string",
                    "description": "SHA3 of the uncles data in the block (32 bytes)"
                },
                "logsBloom": {
                    "type": "string",
                    "description": "the bloom filter for the logs of the block (256 bytes)"
                },
                "transactionsRoot": {
                    "type": "string",
                    "description": "the root of the transaction trie of the block (32 bytes)"
                },
                "stateRoot": {
                    "type": "string",
                    "description": "the root of the final state trie of the block (32 bytes)"
                },
                "miner": {
                    "type": "string",
                    "description": "the address of the beneficiary to whom the mining rewards were given (20 bytes)"
                },
                "difficulty": {
                    "type": "integer",
                    "format": "int64",
                    "description": "integer of the difficulty of this block"
                },
                "totalDifficulty": {
                    "type": "integer",
                    "format": "int64",
                    "description": "integer of the total difficulty of the chain until this block"
                },
                "extraData": {
                    "type": "string",
                    "description": "the \"extra data\" field for this block"
                },
                "gasLimit": {
                    "type": "integer",
                    "format": "int32",
                    "description": "the maximum gas allowed in this block"
                },
                "gasUsed": {
                    "type": "integer",
                    "format": "int32",
                    "description": "the total gas used by all transactions in this block"
                },
                "timestamp": {
                    "type": "integer",
                    "format": "int32",
                    "description": "the unix timestamp for when the block was collated"
                },
                "transactions": {
                    "type": "array",
                    "description": "Array of transaction hashes",
                    "items": {
                        "type": "string"
                    }
                },
                "uncles": {
                    "type": "array",
                    "description": "Array of uncle hashes",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}